@@ -1,13 +1,16 @@
 import ChatScreen from './ChatScreen/ChatScreen';
+import Chats from './ChatScreen/Chats';
+
 import SideBar from './SideBar/SideBar';
 
 const Main = () => {
   return (
-    <section className='flex justify-center items-center h-screen'>
+    <section className='flex justify-center items-center w-full h-screen'>
       <div className='flex ring flex-col'>
         <SideBar />
+
       </div>
-      <ChatScreen />
+      <Chats />
 
 
     </section>
